Add files via upload
GuessingNumber App-
It works on simply if-else condition checking statement and create a Random class which generate Random number 
and user enter the number and app give response like a number is high that random number and low then match the user number 
and random number will be matched it simply message that you got! try again.so that is simply Guessing number.